+++ android/src/test/java/com/instabug/flutter/InstabugApiTest.java
@@ -21,6 +21,7 @@ import android.app.Application;
 import android.graphics.Bitmap;
 import android.net.Uri;
 
+import com.instabug.apm.InternalAPM;
 import com.instabug.bug.BugReporting;
 import com.instabug.flutter.generated.InstabugPigeon;
 import com.instabug.flutter.modules.InstabugApi;
@@ -36,9 +37,14 @@ import com.instabug.library.Platform;
 import com.instabug.library.ReproConfigurations;
 import com.instabug.library.ReproMode;
 import com.instabug.library.featuresflags.model.IBGFeatureFlag;
+import com.instabug.library.internal.crossplatform.CoreFeature;
+import com.instabug.library.internal.crossplatform.FeaturesStateListener;
+import com.instabug.library.internal.crossplatform.InternalCore;
 import com.instabug.library.invocation.InstabugInvocationEvent;
 import com.instabug.library.model.NetworkLog;
 import com.instabug.library.ui.onboarding.WelcomeMessage;
+import com.instabug.survey.Surveys;
+import com.instabug.survey.callbacks.OnShowCallback;
 
 import org.json.JSONObject;
 import org.junit.After;
@@ -59,6 +65,11 @@ import java.util.Map;
 import java.util.concurrent.Callable;
 
 import io.flutter.plugin.common.BinaryMessenger;
+import kotlin.jvm.functions.Function1;
+
+import org.mockito.Mockito;
+import org.mockito.invocation.InvocationOnMock;
+import org.mockito.stubbing.Answer;
 import org.mockito.verification.VerificationMode;
 
 public class InstabugApiTest {
@@ -73,7 +84,10 @@ public class InstabugApiTest {
     @Before
     public void setUp() throws NoSuchMethodException {
         mCustomTextPlaceHolder = mockConstruction(InstabugCustomTextPlaceHolder.class);
-        api = spy(new InstabugApi(mContext, screenshotProvider));
+
+        BinaryMessenger mMessenger = mock(BinaryMessenger.class);
+        final InstabugPigeon.FeatureFlagsFlutterApi flutterApi = new InstabugPigeon.FeatureFlagsFlutterApi(mMessenger);
+        api = spy(new InstabugApi(mContext, screenshotProvider, flutterApi));
         mInstabug = mockStatic(Instabug.class);
         mBugReporting = mockStatic(BugReporting.class);
         mHostApi = mockStatic(InstabugPigeon.InstabugHostApi.class);
@@ -87,6 +101,7 @@ public class InstabugApiTest {
         mBugReporting.close();
         mHostApi.close();
         GlobalMocks.close();
+
     }
 
     @Test
@@ -349,11 +364,11 @@ public class InstabugApiTest {
 
     @Test
     public void testAddFeatureFlags() {
-       Map<String,String > featureFlags = new HashMap<>();
-        featureFlags.put("key1","variant1");
+        Map<String, String> featureFlags = new HashMap<>();
+        featureFlags.put("key1", "variant1");
         api.addFeatureFlags(featureFlags);
-        List<IBGFeatureFlag> flags=new ArrayList<IBGFeatureFlag>();
-        flags.add(new IBGFeatureFlag("key1","variant1"));
+        List<IBGFeatureFlag> flags = new ArrayList<IBGFeatureFlag>();
+        flags.add(new IBGFeatureFlag("key1", "variant1"));
         mInstabug.verify(() -> Instabug.addFeatureFlags(flags));
     }
 
@@ -597,5 +612,22 @@ public class InstabugApiTest {
     public void testWillRedirectToStore() {
         api.willRedirectToStore();
         mInstabug.verify(Instabug::willRedirectToStore);
+    }
+
+
+    @Test
+    public void isW3CFeatureFlagsEnabled() {
+      MockedStatic<InternalCore> internalCoreMockedStatic=  mockStatic(InternalCore.class);
+        Boolean isW3cExternalGeneratedHeaderEnabled= InternalCore.INSTANCE._isFeatureEnabled(CoreFeature.W3C_ATTACHING_GENERATED_HEADER);
+        Boolean isW3cExternalTraceIDEnabled= InternalCore.INSTANCE._isFeatureEnabled(CoreFeature.W3C_EXTERNAL_TRACE_ID);
+        Boolean isW3cCaughtHeaderEnabled= InternalCore.INSTANCE._isFeatureEnabled(CoreFeature.W3C_ATTACHING_CAPTURED_HEADER);
+
+
+        Map<String, Boolean> flags = api.isW3CFeatureFlagsEnabled();
+        assertEquals(isW3cExternalGeneratedHeaderEnabled, flags.get("isW3cExternalGeneratedHeaderEnabled"));
+        assertEquals(isW3cExternalTraceIDEnabled, flags.get("isW3cExternalTraceIDEnabled"));
+        assertEquals(isW3cCaughtHeaderEnabled, flags.get("isW3cCaughtHeaderEnabled"));
+        internalCoreMockedStatic.close();
+
     }
 }