+++ example/ios/InstabugTests/Util/IBGNetworkLogger+Test.h
@@ -18,5 +18,10 @@
                    startTime:(int64_t)startTime
                     duration:(int64_t) duration
                 gqlQueryName:(NSString *_Nullable)gqlQueryName
-          serverErrorMessage:(NSString *_Nullable)gqlServerError;
+          serverErrorMessage:(NSString *_Nullable)gqlServerError
+               isW3cCaughted:(NSNumber *_Nullable)isW3cCaughted
+                   partialID:(NSNumber *_Nullable)partialId
+                   timestamp:(NSNumber *_Nullable)timestamp
+     generatedW3CTraceparent:(NSString *_Nullable)generatedW3CTraceparent
+      caughtedW3CTraceparent:(NSString *_Nullable)caughtedW3CTraceparent;
 @end