+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add support for masking screen names captured by Instabug through the `Instabug.setScreenNameMaskingCallback` API ([#500](https://github.com/Instabug/Instabug-Flutter/pull/500)).
+- Add support for tracing network requests from Instabug to services like Datadog and New Relic ([#481](https://github.com/Instabug/Instabug-Flutter/pull/481)).
 
 ### Changed
 
@@ -39,7 +40,7 @@
 ### Added
 
 - Add support for capturing screen loading time in APM ([#477](https://github.com/Instabug/Instabug-Flutter/pull/477)).
-
+- 
 ### Changed
 
 - Bump Instabug Android SDK to v13.2.0 ([#482](https://github.com/Instabug/Instabug-Flutter/pull/482)). [See release notes](https://github.com/Instabug/Instabug-Android/releases/tag/v13.2.0).